Conflicts:
	Cryptosystem/src/keyGen/KeyTuple.java
	Cryptosystem/src/keyGen/TrapdoorSampler.java

+++ Cryptosystem/src/encrypt/Encrypter.java
@@ -1,7 +1,19 @@
 package encrypt;
 
+
+import java.math.BigInteger;
+
+/**
+ * 
+ * @author Aaron
+ * Contains method to encrypt plaintext and related helper-methods.
+ *
+ */
+
+
+
 public interface Encrypter {
 
-	public long[][] encrypt(long[][] plaintext ,long[][] publicKey);
+	public BigInteger[][] encrypt(BigInteger[][] plaintext ,BigInteger[][] publicKey);
 	
 }
