+++ Cryptosystem/src/homomorphicOperations/HomomorphicOperations.java
@@ -1,6 +1,15 @@
 package homomorphicOperations;
 
+import java.math.BigInteger;
+
 public interface HomomorphicOperations {
-	public long[][] add(long[][] c1, long[][] c2);
-	public long[][] multiply(long[][] c1,long[][] c2);
+	/**
+	 * adds two ciphertexts
+	 */
+	public BigInteger[][] add(BigInteger[][] c1, BigInteger[][] c2);
+	/**
+	 * multiplies two ciphertexts
+	 */
+	public BigInteger[][] multiply(BigInteger[][] c1,BigInteger[][] c2);
+
 }
